+++ GRDB Tests/Public/Record/MinimalPrimaryKeyRowIDTests.swift
@@ -256,8 +256,8 @@ class MinimalPrimaryKeyRowIDTests: GRDBTestCase {
             try dbQueue.inDatabase { db in
                 let record = MinimalRowID()
                 record.id = 123456
-                let deletionResult = try record.delete(db)
-                XCTAssertEqual(deletionResult, Record.DeletionResult.NoRowDeleted)
+                let changes = try record.delete(db)
+                XCTAssertEqual(changes.changedRowCount, 0)
             }
         }
     }
@@ -267,8 +267,8 @@ class MinimalPrimaryKeyRowIDTests: GRDBTestCase {
             try dbQueue.inDatabase { db in
                 let record = MinimalRowID()
                 try record.insert(db)
-                let deletionResult = try record.delete(db)
-                XCTAssertEqual(deletionResult, Record.DeletionResult.RowDeleted)
+                let changes = try record.delete(db)
+                XCTAssertEqual(changes.changedRowCount, 1)
                 
                 let row = Row.fetchOne(db, "SELECT * FROM minimalRowIDs WHERE id = ?", arguments: [record.id])
                 XCTAssertTrue(row == nil)
@@ -281,10 +281,10 @@ class MinimalPrimaryKeyRowIDTests: GRDBTestCase {
             try dbQueue.inDatabase { db in
                 let record = MinimalRowID()
                 try record.insert(db)
-                var deletionResult = try record.delete(db)
-                XCTAssertEqual(deletionResult, Record.DeletionResult.RowDeleted)
-                deletionResult = try record.delete(db)
-                XCTAssertEqual(deletionResult, Record.DeletionResult.NoRowDeleted)
+                var changes = try record.delete(db)
+                XCTAssertEqual(changes.changedRowCount, 1)
+                changes = try record.delete(db)
+                XCTAssertEqual(changes.changedRowCount, 0)
             }
         }
     }