+++ GRDB Tests/Public/Record/PrimaryKeyMultipleTests.swift
@@ -268,8 +268,8 @@ class PrimaryKeyMultipleTests: GRDBTestCase {
         assertNoError {
             try dbQueue.inDatabase { db in
                 let record = Citizenship(personName: "Arthur", countryName: "France", native: true)
-                let deletionResult = try record.delete(db)
-                XCTAssertEqual(deletionResult, Record.DeletionResult.NoRowDeleted)
+                let changes = try record.delete(db)
+                XCTAssertEqual(changes.changedRowCount, 0)
             }
         }
     }
@@ -279,8 +279,8 @@ class PrimaryKeyMultipleTests: GRDBTestCase {
             try dbQueue.inDatabase { db in
                 let record = Citizenship(personName: "Arthur", countryName: "France", native: true)
                 try record.insert(db)
-                let deletionResult = try record.delete(db)
-                XCTAssertEqual(deletionResult, Record.DeletionResult.RowDeleted)
+                let changes = try record.delete(db)
+                XCTAssertEqual(changes.changedRowCount, 1)
                 
                 let row = Row.fetchOne(db, "SELECT * FROM citizenships WHERE personName = ? AND countryName = ?", arguments: [record.personName, record.countryName])
                 XCTAssertTrue(row == nil)
@@ -293,10 +293,10 @@ class PrimaryKeyMultipleTests: GRDBTestCase {
             try dbQueue.inDatabase { db in
                 let record = Citizenship(personName: "Arthur", countryName: "France", native: true)
                 try record.insert(db)
-                var deletionResult = try record.delete(db)
-                XCTAssertEqual(deletionResult, Record.DeletionResult.RowDeleted)
-                deletionResult = try record.delete(db)
-                XCTAssertEqual(deletionResult, Record.DeletionResult.NoRowDeleted)
+                var changes = try record.delete(db)
+                XCTAssertEqual(changes.changedRowCount, 1)
+                changes = try record.delete(db)
+                XCTAssertEqual(changes.changedRowCount, 0)
             }
         }
     }